+++ src/renderer/src/bootstrap.test.ts
@@ -33,6 +33,7 @@ function noopDeps(): CommandDeps {
     toggleMaximize: () => {},
     toggleRail: () => {},
     openPalette: () => {},
+    openFind: () => {},
     toggleTabRows: () => {},
     jumpToNextAttention: () => {},
     openAbout: () => {},